+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Bytes.split_at`
+- `Vec.split_at`
+- `Iteratr.chunks`
+
+### Added
+
 - `Bytes.zeroed`
 - `Bytes.to_vec`
 - `Bytes.fill`